Move rollout reconstruction tests out of codex.rs
